+++ source/vulkan/ui/ui_widget.d
@@ -120,7 +120,10 @@ abstract class UiWidget
         for (ptrdiff_t index = cast(ptrdiff_t)children.length - 1; index >= 0; --index)
         {
             if (children[cast(size_t)index].dispatchPointerEvent(childEvent))
+            {
+                event.actionActivated = event.actionActivated || childEvent.actionActivated;
                 return true;
+            }
         }
 
         return handlePointerEvent(event);